+++ crates/draff/Cargo.toml
@@ -8,4 +8,20 @@ edition = "2018"
 clap = { version = "2.33.0", features = ["yaml"] }
 env_logger = "0.7.1"
 exitfailure = "0.5.1"
+failure = "0.1.6"
+failure_derive = "0.1.6"
+lazy-init = "0.3.0"
+log = "0.4.8"
+md5 = "0.7.0"
+itertools = "0.8.2"
+pdatastructs = { git = "https://github.com/luizirber/pdatastructs.rs", branch = "succinct_wasm" }
+serde = "1.0.103"
+serde_derive = "1.0.103"
+serde_json = "1.0.44"
 sourmash = { path = "../sourmash/" }
+ukhs = { git = "https://github.com/luizirber/ukhs", branch = "feature/alternative_backends", features = ["boomphf_mphf"], default-features = false}
+
+[dependencies.needletail]
+version = "0.3.2"
+default-features = false
+#features = ["compression"]
