+++ crates/draff/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod cmd;
+pub mod index;
+pub mod sketch;
