+++ crates/sourmash/Cargo.toml
@@ -36,15 +36,8 @@ once_cell = "1.2.0"
 serde = "1.0.103"
 serde_derive = "1.0.103"
 serde_json = "1.0.44"
-#ukhs = { git = "https://github.com/luizirber/ukhs", branch = "feature/alternative_backends", features = ["boomphf_mphf"], default-features = false}
 primal = "0.2.3"
-#pdatastructs = { git = "https://github.com/luizirber/pdatastructs.rs", branch = "succinct_wasm" }
 typed-builder = "0.4.0"
-
-[dependencies.needletail]
-version = "0.3.2"
-default-features = false
-#features = ["compression"]
 
 [target.'cfg(all(target_arch = "wasm32", target_vendor="unknown"))'.dependencies.wasm-bindgen]
 version = "0.2.55"
@@ -64,6 +57,11 @@ tempfile = "3.1.0"
 assert_matches = "1.3.0"
 assert_cmd = "0.12.0"
 predicates = "1.0.2"
+
+[dev-dependencies.needletail]
+version = "0.3.2"
+default-features = false
+#features = ["compression"]
 
 [[bench]]
 name = "index"
